Code refactor, linters fix
